server/requirements.txt
fastapi
uvicorn
chromadb
langchain
langgraph
langchain-openai
openai
pydantic
python-dotenv
bs4